Update ShippingEasy/OrderQuery.cs
Orders query replace O with u. .NET. That is the root of the problem - the space is getting escaped differently on the client than in our server. As a result, our server calculates a different HMAC value and when the value the server calculates does not match what the client calculated and specified in the api_signature URL parameter, the end result is the 401 error you are getting.